Merge 0.24

See merge request hzero-c7ncd/devops-service!1506

+++ src/main/java/io/choerodon/devops/app/service/GitlabUserService.java
@@ -20,9 +20,9 @@ public interface GitlabUserService {
 
     void updateGitlabUser(GitlabUserRequestVO gitlabUserReqDTO);
 
-    void isEnabledGitlabUser(Integer userId);
+    void isEnabledGitlabUser(Long userId);
 
-    void disEnabledGitlabUser(Integer userId);
+    void disEnabledGitlabUser(Long userId);
 
     Boolean doesEmailExists(String email);
