+++ src/main/java/io/choerodon/devops/app/service/impl/GitlabUserServiceImpl.java
@@ -80,7 +80,7 @@ public class GitlabUserServiceImpl implements GitlabUserService {
     }
 
     @Override
-    public void isEnabledGitlabUser(Integer userId) {
+    public void isEnabledGitlabUser(Long userId) {
         UserAttrDTO userAttrDTO = userAttrService.baseQueryById(TypeUtil.objToLong(userId));
         if (userAttrDTO != null) {
             gitlabServiceClientOperator.enableUser(TypeUtil.objToInteger(userAttrDTO.getGitlabUserId()));
@@ -88,7 +88,7 @@ public class GitlabUserServiceImpl implements GitlabUserService {
     }
 
     @Override
-    public void disEnabledGitlabUser(Integer userId) {
+    public void disEnabledGitlabUser(Long userId) {
         UserAttrDTO userAttrDTO = userAttrService.baseQueryById(TypeUtil.objToLong(userId));
         if (userAttrDTO != null) {
             gitlabServiceClientOperator.disableUser(TypeUtil.objToInteger(userAttrDTO.getGitlabUserId()));
